Print heredocs consistently
