+++ src/nodes/arrays.js
@@ -5,7 +5,6 @@ const {
   indent,
   join,
   line,
-  literalline,
   softline
 } = require("../prettier");
 
@@ -43,25 +42,6 @@ const printSpecialArray = (parts) =>
     ])
   );
 
-// Extract out the actual elements, taking into account nesting with
-// `args_add_star` nodes. The only nodes that get passed into this function are
-// `args` or `args_add_star`.
-const getElements = (node, elementPath) => {
-  if (node.type === "args") {
-    return node.body.map((element, index) => ({
-      element,
-      elementPath: elementPath.concat(["body", index])
-    }));
-  }
-
-  return getElements(node.body[0], elementPath.concat(["body", 0])).concat(
-    node.body.slice(1).map((element, index) => ({
-      element,
-      elementPath: elementPath.concat(["body", index + 1])
-    }))
-  );
-};
-
 module.exports = {
   array: (path, { addTrailingCommas }, print) => {
     const args = path.getValue().body[0];
@@ -86,44 +66,18 @@ module.exports = {
       return printSpecialArray(path.call(print, "body", 0));
     }
 
-    const normalDocs = [];
-
-    const elementDocs = path.call(print, "body", 0);
-    const elements = getElements(path.getValue().body[0], ["body", 0]);
-
-    // We need to manually loop through the elements in the array in order to
-    // take care of heredocs printing (their commas go after the opening, as
-    // opposed to at the end).
-    elements.forEach(({ element, elementPath }, index) => {
-      const isInner = index !== elements.length - 1;
-
-      if (element.type === "heredoc") {
-        normalDocs.push(
-          element.beging,
-          isInner || addTrailingCommas ? "," : "",
-          literalline,
-          concat(
-            path.map.apply(path, [print].concat(elementPath).concat("body"))
-          ),
-          element.ending,
-          isInner ? line : ""
-        );
-      } else {
-        normalDocs.push(elementDocs[index]);
-
-        if (isInner) {
-          normalDocs.push(concat([",", line]));
-        } else if (addTrailingCommas) {
-          normalDocs.push(ifBreak(",", ""));
-        }
-      }
-    });
-
     return group(
       concat([
         "[",
-        indent(concat([softline].concat(normalDocs))),
-        concat([softline, "]"])
+        indent(
+          concat([
+            softline,
+            join(concat([",", line]), path.call(print, "body", 0)),
+            addTrailingCommas ? ifBreak(",", "") : ""
+          ])
+        ),
+        softline,
+        "]"
       ])
     );
   },
