+++ src/nodes/hashes.js
@@ -1,12 +1,4 @@
-const {
-  concat,
-  group,
-  ifBreak,
-  indent,
-  join,
-  line,
-  literalline
-} = require("../prettier");
+const { concat, group, ifBreak, indent, join, line } = require("../prettier");
 
 const { nodeDive, prefix, skipAssignIndent } = require("../utils");
 
@@ -20,12 +12,12 @@ const { nodeDive, prefix, skipAssignIndent } = require("../utils");
 //
 // This function represents that check, as it determines if it can convert the
 // symbol node into a hash label.
-const isValidHashLabel = (symbolLiteral) => {
+function isValidHashLabel(symbolLiteral) {
   const label = symbolLiteral.body[0].body[0].body;
   return label.match(/^[_A-Za-z]/) && !label.endsWith("=");
-};
+}
 
-const makeLabel = (path, { preferHashLabels }, print, steps) => {
+function makeLabel(path, { preferHashLabels }, print, steps) {
   const labelNode = nodeDive(path.getValue(), steps);
   const labelDoc = path.call.apply(path, [print].concat(steps));
 
@@ -54,7 +46,20 @@ const makeLabel = (path, { preferHashLabels }, print, steps) => {
     default:
       return concat([labelDoc, " =>"]);
   }
-};
+}
+
+function printAssocNew(path, opts, print) {
+  const valueDoc = path.call(print, "body", 1);
+  const parts = [makeLabel(path, opts, print, ["body", 0])];
+
+  if (skipAssignIndent(path.getValue().body[1])) {
+    parts.push(" ", valueDoc);
+  } else {
+    parts.push(indent(concat([line, valueDoc])));
+  }
+
+  return group(concat(parts));
+}
 
 function printHash(path, { addTrailingCommas }, print) {
   const hashNode = path.getValue();
@@ -71,18 +76,10 @@ function printHash(path, { addTrailingCommas }, print) {
   // wrapping it in an array in case we need to append a trailing comma.
   const assocDocs = [path.call(print, "body", 0)];
 
-  // Here we get a reference to the last key-value pair's value node, in order
-  // to check if we're dealing with a heredoc. If we are, then the trailing
-  // comma printing is handled from within the assoclist_from_args node
-  // printing, because the trailing comma has to go after the heredoc
-  // declaration.
-  const assocNodes = hashNode.body[0].body[0];
-  const lastAssocValueNode = assocNodes[assocNodes.length - 1].body[1];
-
   // If we're adding a trailing comma and the last key-value pair's value node
   // is not a heredoc node, then we can safely append the extra comma if the
   // hash ends up getting printed on multiple lines.
-  if (addTrailingCommas && lastAssocValueNode.type !== "heredoc") {
+  if (addTrailingCommas) {
     assocDocs.push(ifBreak(",", ""));
   }
 
@@ -95,53 +92,14 @@ function printHash(path, { addTrailingCommas }, print) {
   );
 }
 
+function printHashContents(path, opts, print) {
+  return group(join(concat([",", line]), path.map(print, "body", 0)));
+}
+
 module.exports = {
-  assoc_new: (path, opts, print) => {
-    const valueDoc = path.call(print, "body", 1);
-    const parts = [makeLabel(path, opts, print, ["body", 0])];
-
-    if (skipAssignIndent(path.getValue().body[1])) {
-      parts.push(" ", valueDoc);
-    } else {
-      parts.push(indent(concat([line, valueDoc])));
-    }
-
-    return group(concat(parts));
-  },
+  assoc_new: printAssocNew,
   assoc_splat: prefix("**"),
-  assoclist_from_args: (path, opts, print) => {
-    const { addTrailingCommas } = opts;
-
-    const assocNodes = path.getValue().body[0];
-    const assocDocs = [];
-
-    assocNodes.forEach((assocNode, index) => {
-      const isInner = index !== assocNodes.length - 1;
-      const valueNode = assocNode.body[1];
-
-      if (valueNode && valueNode.type === "heredoc") {
-        assocDocs.push(
-          makeLabel(path, opts, print, ["body", 0, index, "body", 0]),
-          " ",
-          valueNode.beging,
-          isInner || addTrailingCommas ? "," : "",
-          literalline,
-          concat(path.map(print, "body", 0, index, "body", 1, "body")),
-          valueNode.ending,
-          isInner ? line : ""
-        );
-      } else {
-        assocDocs.push(path.call(print, "body", 0, index));
-
-        if (isInner) {
-          assocDocs.push(concat([",", line]));
-        }
-      }
-    });
-
-    return group(concat(assocDocs));
-  },
-  bare_assoc_hash: (path, opts, print) =>
-    group(join(concat([",", line]), path.map(print, "body", 0))),
+  assoclist_from_args: printHashContents,
+  bare_assoc_hash: printHashContents,
   hash: printHash
 };
